Bugfix in _fix_gnuplot_svg_size and extra tp arguments
The width/height were not correctly read in the _fix_gnuplot_svg_size
function - they were strings causing svg.setAttribute calls to crash
when '%dpx' was used. Also, if no args.size are provided, the picture
size needs to be passed as None to this function, or the default size
variables are used.

An additional parameter -tp has been added (also to the --help) to
allow arbitrary parameters to be passed to the tikzpicture call.