Added flag to scale svg images in the notebook view
The svg options have been fixed in previous commits
such that they default to their natural canvas values.
The image may appear large or small in the notebook
view, however. An additional flag -svgsc allows scaling
of the final image in the notebook view.

Note that this is not the same as the -sc option,
which does not scale the whole completed image.
The -sc option applies scaling to tikz picture,
which in turn applies different behaviour to text and
nodes etc., i.e. the relative size of text and lines
change.